requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
dataclasses-json==0.6.6
Deprecated==1.2.14
dune_client==1.7.3
et-xmlfile==1.1.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
ndjson==0.3.1
numpy==1.26.4
openpyxl==3.1.3
packaging==24.0
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.3
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
stripe==9.12.0
types-Deprecated==1.2.9.20240311
types-python-dateutil==2.9.0.20240316
types-PyYAML==6.0.12.20240311
types-requests==2.32.0.20240602
types-setuptools==70.0.0.20240524
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
